move server/utils.js to server/routes/utils.js
